Make Diagnostics a submodule
