+++ src/ClimaAtmos.jl
@@ -110,14 +110,13 @@ include(joinpath("prognostic_equations", "limited_tendencies.jl"))
 
 include(joinpath("callbacks", "callbacks.jl"))
 
+include(joinpath("diagnostics", "Diagnostics.jl"))
+
 include(joinpath("solver", "model_getters.jl")) # high-level (using parsed_args) model getters
 include(joinpath("solver", "type_getters.jl"))
 include(joinpath("solver", "yaml_helper.jl"))
 include(joinpath("solver", "solve.jl"))
 
 include(joinpath("parameters", "create_parameters.jl"))
-
-include(joinpath("diagnostics", "Diagnostics.jl"))
-include(joinpath("diagnostics", "Writers.jl"))
 
 end # module
